+++ src/ru/ifmo/ctddev/komarov/calc/parsetree/NodeNumber.java
@@ -15,7 +15,7 @@ public class NodeNumber implements ParseTreeNode {
     }
 
     @Override
-    public double evaluate(int x) {
+    public double evaluate(double x) {
         return value;
     }
 }
